Added bypass to non-coded fight scene
Doesn't work if hit monster when leaving entrance room for first time. Also, because of the lazy way I did it, there is a visual "glitch" when the player returns to the dungeon: the ball first exits the exit it took into the the first room, then rolls out of the exit of the new room correctly. Would need significant refactoring to fix. Not worth it for extra time I shouldn't be spending on this!